Perms

+++ routes/web.php
@@ -19,9 +19,9 @@ Route::get('/search', 'PostController@search')->name('posts.search');
 
 // THEMES
 Route::get('/', 'ThemeController@index')->name('themes.index');
-Route::get('/follow/{id}', 'ThemeController@follow')->name('themes.follow');
-Route::get('/unfollow/{id}', 'ThemeController@unfollow')->name('themes.unfollow');
-Route::get('/following', 'ThemeController@following')->name('themes.following');
+Route::get('/follow/{id}', 'ThemeController@follow')->name('themes.follow')->middleware('can:admin-user');
+Route::get('/unfollow/{id}', 'ThemeController@unfollow')->name('themes.unfollow')->middleware('can:admin-user');
+Route::get('/following', 'ThemeController@following')->name('themes.following')->middleware('can:admin-user');
 
 // Index sorting
 Route::get('/post_asc', 'ThemeController@indexPostAsc', 'var')->name('themes.indexPostAsc');
@@ -33,14 +33,14 @@ Route::get('/monpost_desc', 'ThemeController@indexMonPostDesc', 'var')->name('th
 Route::get('/view_desc', 'ThemeController@indexViewDesc', 'var')->name('themes.indexViewDesc');
 Route::get('/monview_desc', 'ThemeController@indexMonViewDesc', 'var')->name('themes.indexMonViewDesc');
 
-Route::get('/theme/create', 'ThemeController@create')->name('themes.create');
-Route::post('/', 'ThemeController@store')->name('themes.store');
+Route::get('/theme/create', 'ThemeController@create')->name('themes.create')->middleware('can:admin');
+Route::post('/', 'ThemeController@store')->name('themes.store')->middleware('can:admin');
 Route::get('/{id}', 'ThemeController@show')->name('themes.show');
-Route::get('/theme/edit/{id}', 'ThemeController@edit')->name('themes.edit');
-Route::put('/lock/{id}', 'ThemeController@lock')->name('themes.lock');
-Route::put('/unlock/{id}', 'ThemeController@unlock')->name('themes.unlock');
-Route::put('/theme/{id}', 'ThemeController@update')->name('themes.update');
-Route::delete('/theme/{id}', 'ThemeController@destroy')->name('themes.destroy');
+Route::get('/theme/edit/{id}', 'ThemeController@edit')->name('themes.edit')->middleware('can:admin');
+Route::put('/lock/{id}', 'ThemeController@lock')->name('themes.lock')->middleware('can:admin');
+Route::put('/unlock/{id}', 'ThemeController@unlock')->name('themes.unlock')->middleware('can:admin');
+Route::put('/theme/{id}', 'ThemeController@update')->name('themes.update')->middleware('can:admin');
+Route::delete('/theme/{id}', 'ThemeController@destroy')->name('themes.destroy')->middleware('can:admin');
 
 // POSTS
 // Index sorting
@@ -55,26 +55,25 @@ Route::get('/theme/{id}/moncom_desc', 'ThemeController@showMonComDesc', 'var')->
 Route::get('/theme/{id}/view_desc', 'ThemeController@showViewDesc', 'var')->name('themes.showViewDesc');
 Route::get('/theme/{id}/monview_desc', 'ThemeController@showMonViewDesc', 'var')->name('themes.showMonViewDesc');
 
-Route::get('/post/create', 'PostController@create')->name('posts.create');
-Route::post('/post/create', 'PostController@store')->name('posts.store');
+Route::get('/post/create', 'PostController@create')->name('posts.create')->middleware('can:admin-user');
+Route::post('/post/create', 'PostController@store')->name('posts.store')->middleware('can:admin-user');
 Route::get('/post/{id}', 'PostController@show')->name('posts.show');
-Route::get('/post/{id}/edit', 'PostController@edit')->name('posts.edit');
-Route::put('/post/{id}', 'PostController@update')->name('posts.update');
-Route::delete('/post/{id}', 'PostController@destroy')->name('posts.destroy');
+Route::get('/post/{id}/edit', 'PostController@edit')->name('posts.edit')->middleware('can:admin-user');
+Route::put('/post/{id}', 'PostController@update')->name('posts.update')->middleware('can:admin-user');
+Route::delete('/post/{id}', 'PostController@destroy')->name('posts.destroy')->middleware('can:admin-user');
 
 // USERS
 Route::get('/profile/{id}', 'UserController@show')->name('users.show');
-Route::get('/user/profile', 'UserController@profile')->name('users.profile');
-Route::get('/user/follow/{id}', 'UserController@follow')->name('users.follow');
-Route::get('/user/unfollow/{id}', 'UserController@unfollow')->name('users.unfollow');
-Route::get('/user/following', 'UserController@following')->name('users.following');
+Route::get('/user/follow/{id}', 'UserController@follow')->name('users.follow')->middleware('can:admin-user');
+Route::get('/user/unfollow/{id}', 'UserController@unfollow')->name('users.unfollow')->middleware('can:admin-user');
+Route::get('/user/following', 'UserController@following')->name('users.following')->middleware('can:admin-user');
 
 // COMMENTS
-Route::post('/comment/{id}', 'CommentController@store')->name('comments.store');
-Route::delete('/comment/{id}', 'CommentController@destroy')->name('comments.destroy');
-Route::get('/comment/{id}/edit', 'CommentController@edit')->name('comments.edit');
-Route::put('/comment/{id}', 'CommentController@update')->name('comments.update');
+Route::post('/comment/{id}', 'CommentController@store')->name('comments.store')->middleware('can:admin-user');
+Route::delete('/comment/{id}', 'CommentController@destroy')->name('comments.destroy')->middleware('can:admin-user');
+Route::get('/comment/{id}/edit', 'CommentController@edit')->name('comments.edit')->middleware('can:admin-user');
+Route::put('/comment/{id}', 'CommentController@update')->name('comments.update')->middleware('can:admin-user');
 
 // WARNINGS
-Route::get('/profile/{id}/warn', 'WarningController@create')->name('warnings.create');
-Route::post('/profile/{id}/warn', 'WarningController@store')->name('warnings.store');
+Route::get('/profile/{id}/warn', 'WarningController@create')->name('warnings.create')->middleware('can:admin');
+Route::post('/profile/{id}/warn', 'WarningController@store')->name('warnings.store')->middleware('can:admin');
